Review points: remove SimdOps and NativeSimdOps, move logic to VectorUtilSupport
