+++ jvector-native/src/main/java/io/github/jbellis/jvector/vector/NativeVectorizationProvider.java
@@ -38,7 +38,7 @@ public class NativeVectorizationProvider extends VectorizationProvider {
         if (!NativeSimdOps.check_compatibility()) {
             throw new UnsupportedOperationException("Native SIMD operations are not supported on this platform due to missing CPU support.");
         }
-        this.vectorUtilSupport = new NativeVectorUtilSupport(new VectorSimdOps());
+        this.vectorUtilSupport = new NativeVectorUtilSupport();
         this.vectorTypeSupport = new MemorySegmentVectorProvider();
     }
